Fix Db query syntax. Add/Remove of dealers, products done.

+++ app/models/relations/DealerProduct.scala
@@ -3,5 +3,5 @@ package models.relations
 /**
  * Created by ruraj on 7/17/14.
  */
-case class DealerProduct(dealerId: Int = 0, products: Set[models.Product])
+case class DealerProduct(dealerId: Long = 0, products: Set[models.Product])
